added save and converted double values to floats

+++ src/include/voxel.h
@@ -7,14 +7,14 @@ class Voxel
 {
 public:
     Voxel( void );
-    Voxel( double value );
-    Voxel( double value, bool isBoundary );
+    Voxel( float value );
+    Voxel( float value, bool isBoundary );
     bool isBoundary( void );
-    void setValue( double value );
-    double getValue( void );
+    void setValue( float value );
+    float getValue( void );
 
 private:
-    double _value;
+    float _value;
     bool _isBoundary;
 };
